api/requirements.txt
hume==0.8.6
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
openai
audioop-lts; python_version >= "3.13"
pyjwt
passlib[bcrypt]
sqlalchemy>=2.0.0
psycopg2-binary
alembic

